Add files via upload
Trying to make an admin page to view/edit/delete/create anything in my mysql table. I keep getting an error : "Call to undefined function mysql_query() in /Applications/XAMPP/xamppfiles/htdocs/pet-shop/admin/list.php:20 Stack trace: #0 {main} thrown in"

When I use list.php, it is supposed to show an option to delete and edit (using edit.php and delete.php) but of course that isn't working because of the error.